[BUGFIX] Add process function, move recover into new function
The receiver is an infinite loop that reads from a socket, unless
signalled for close. If we recover that function, we are outside the
loop and will not process anymore data.

This commit creates a new function to be called every time new data
is consumed, therefore keeping the "receiver" goroutine in its loop.